Check for locally committed changes
This change is for the case of deployment from individual
systems. It's important to check to ensure that there are no
locally committed changes in a tag, otherwise a checkout on the
minions will fail since the referenced object won't exist.

This change adds two new configuration options:

  deploy.allow-local-commits:
    Required: False
    Default: True

  deploy.origin-deploy-branch
    Required: False
    Default: 'origin/master'

If deploy.allow-local-commits is set to false then the sync action
will check to see if there are any local commits that do not exist
in the deploy.origin-deploy-branch, then an error will be reported
and sync will exit without writing a sync tag or starting a sync.